# Using maps for slice keys so that we don't have to worry about the so…