This eliminates slicing errors due to the bytemap or the slice keys not being sorted lexicographically. Interestingly enough, it doesn't seem to hurt slice performance.

Benchmarks before change:

```
goos: darwin
goarch: amd64
pkg: github.com/getlantern/bytemap
BenchmarkNew-6                       	  455803	      2482 ns/op
BenchmarkFromSortedKeysAndValues-6   	 1711551	       701 ns/op
BenchmarkByteMapAllKeys-6            	  229071	      5231 ns/op
BenchmarkByteMapOneKey-6             	  430521	      2691 ns/op
BenchmarkByteSlice-6                 	 1000000	      1031 ns/op
BenchmarkMsgPackAllKeys-6            	  126978	      9500 ns/op
BenchmarkMsgPackOneKey-6             	  294931	      4129 ns/op
BenchmarkMsgPackSlice-6              	  170470	      7088 ns/op
BenchmarkReadKeysIndividually-6      	 4182805	       291 ns/op
BenchmarkReadKeysIteration-6         	 9404410	       124 ns/op
```

Benchmarks after change:

```
goos: darwin
goarch: amd64
pkg: github.com/getlantern/bytemap
BenchmarkNew-6                       	  448096	      2581 ns/op
BenchmarkFromSortedKeysAndValues-6   	 1726626	       698 ns/op
BenchmarkByteMapAllKeys-6            	  218368	      5397 ns/op
BenchmarkByteMapOneKey-6             	  430870	      3347 ns/op
BenchmarkByteSlice-6                 	 1449924	       888 ns/op
BenchmarkMsgPackAllKeys-6            	   93820	     10685 ns/op
BenchmarkMsgPackOneKey-6             	  282267	      4306 ns/op
BenchmarkMsgPackSlice-6              	  153225	      7460 ns/op
BenchmarkReadKeysIndividually-6      	 3659856	       320 ns/op
BenchmarkReadKeysIteration-6         	 9567508	       126 ns/op
```

- [x] Do the tests pass? Consistently?
- [x] Did this change improve test coverage?